feat: add windows legacy GDI capture backends
PrintWindow/BitBlt paths with pump pre-probe, GDI RAII, and fixture-proven pixels so screenshot can fall back when modern capture is unavailable.

+++ crates/windows/src/tree/fixture_pattern_gdi.rs
@@ -254,7 +254,7 @@ impl DibCapture {
                 biHeight: -height,
                 biPlanes: 1,
                 biBitCount: 32,
-                biCompression: BI_RGB as u32,
+                biCompression: BI_RGB,
                 ..Default::default()
             },
             ..Default::default()
